rename variant to lighthouse
