+++ packages/shared/src/variant_map.ts
@@ -17,7 +17,7 @@ import { badukVariant } from "./variants/baduk";
 import { Variant } from "./variant";
 import markdownit from "markdown-it";
 import { sFractionalVariant } from "./variants/s_fractional";
-import { fogOfWarVariant } from "./variants/fog_of_war/fog_of_war";
+import { lighthouseVariant } from "./variants/lighthouse/lighthouse";
 
 const variant_map: {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -39,7 +39,7 @@ const variant_map: {
   drift: driftVariant,
   quantum: quantumVariant,
   sfractional: sFractionalVariant,
-  "fog of war": fogOfWarVariant,
+  lighthouse: lighthouseVariant,
 };
 
 class ConfigError extends Error {
