+++ packages/vue-client/src/board_map.ts
@@ -5,7 +5,7 @@ import KeimaBoard from "@/components/boards/KeimaBoard.vue";
 import type { Component } from "vue";
 import QuantumBoard from "@/components/boards/QuantumBoard.vue";
 import DefaultBoard from "./components/boards/DefaultBoard.vue";
-import FogOfWarBoard from "./components/boards/FogOfWarBoard.vue";
+import LighthouseBoard from "./components/boards/LighthouseBoard.vue";
 
 const board_map: {
   [variant: string]: Component<{ config: unknown; gamestate: unknown }>;
@@ -15,7 +15,7 @@ const board_map: {
   fractional: FractionalBoard,
   keima: KeimaBoard,
   quantum: QuantumBoard,
-  "fog of war": FogOfWarBoard,
+  lighthouse: LighthouseBoard,
 };
 
 export function getBoard(variant: string) {
