+++ packages/vue-client/src/components/boards/LighthouseBoard.vue
